Finally updated the code to build against the latest versions of libnds, libfat, and wifilib.  Many of these changes were renames to use the new register names, however some changes had to be made to accomodate the new initialization code present in libnds.
Additionally, the code has been fixed such that cross-compiling NHDS on
a 64-bit host works correctly (NetHack uses long's in a few places where
it really matters that the host and target platforms end up with the
same widths.  As a result, that code had to be changed to use uint32_t).

Finally, in the process of fixing some keyboard related bugs, I
reformatted and substantially cleaned up the keyboard code so it's
actually readable and comprehensible, as well as clean and consistent.