Fixed Makefile if build_idx doesn't exist.
Technically, the 'tags' target either shouldn't be in the Makefile, or
build_idx should be made sufficiently general to work on machines other
than my laptop.  But the target is really convenient, and I'm lazy, so
instead, I just changed the target so that it'll proceed if build_idx
isn't present. *shrug*