Added support for selecting a # of items in yn_function.  This fixes #1.
There's two principle parts to this change.  First off, I added a new
custom selection mode for menus called PICK_ANY_ONE.  In this mode, you
can select one type of item, and count it up if you want, but you can't
select multiple items.  So it's a kinda of limited PICK_ANY.

I then altered nds_yn_function so that, if the character list passed in
is NULL or contains a '#', then we use PICK_ANY_ONE as the menu mode,
and after an item is selected, we form a string containing the count and
the inventory letter concatenated together.  Finally, we return the
first character of that buffer, and shove the rest into the
newly-refactored input buffer.