Various build fixes.
I'm not sure why this build supposedly worked before (or, more
likely, I fixed this issue in my local tree and then blew away
the tree), but strncmpi isn't available for the DS so we need to
not define STRNCMPI in the build.

While I was in here, I also cleaned things up a built to remove
some unnecessary code here and there.