Updating README to point to new repository
This repository is no longer maintained!  All work on this port has
moved to:

https://github.com/fancypantalons/NetHack

That tree is a true fork of the DevTeam tree with the DS port overlaid.

The DS port is also a much cleaner port with minimal customizations,
while taking advantage of new infrastructure (like status hilites)
available in 3.6.0 and later!